+++ Sources/Pippin/Core/Environment/LaunchArguments.swift
@@ -37,7 +37,8 @@ import Foundation
     /// Upon initializing the app's model layer, remove any entities stored in it.
     case wipeDataModel
     
-    case visualizeTouches
+    /// Upon launch, remove any stored user defaults used by the app.
+    case wipeDefaults
 
     /// Determine if the calling launch argument was activated on the app launch.
     ///
@@ -57,7 +58,7 @@ import Foundation
         case .simulateInAppPurchaseRestoreFailure: return "simulate-in-app-purchase-restore-failure"
         case .wipePersistedInAppPurchases: return "wipe-in-app-purchases"
         case .wipeDataModel: return "wipe-data-model"
-        case .visualizeTouches: return "visualize-touches"
+        case .wipeDefaults: return "wipe-defaults"
         }
     }